Application changes for Franchise Owner Role implementation

+++ DataLayer/sp_GetBhkWiseBookingDetails_Result.cs
@@ -13,7 +13,7 @@ namespace DataLayer
     
     public partial class sp_GetBhkWiseBookingDetails_Result
     {
-        public Nullable<int> Bhk { get; set; }
+        public Nullable<double> Bhk { get; set; }
         public Nullable<int> TotalBookingamount { get; set; }
         public string CreatedDate { get; set; }
         public Nullable<int> Year { get; set; }
