+++ DataLayer/sp_GetFlatDetails_Result.cs
@@ -14,7 +14,7 @@ namespace DataLayer
     public partial class sp_GetFlatDetails_Result
     {
         public Nullable<int> Floor { get; set; }
-        public Nullable<int> Bhk { get; set; }
+        public Nullable<double> Bhk { get; set; }
         public string Facing { get; set; }
         public Nullable<int> Sft { get; set; }
         public Nullable<int> ClubHouseCharges { get; set; }
