+++ ModelLayer/FlatDetails.cs
@@ -9,7 +9,7 @@ namespace ModelLayer
     public partial class FlatDetails
     {
         public Nullable<int> Floor { get; set; }
-        public Nullable<int> Bhk { get; set; }
+        public Nullable<double> Bhk { get; set; }
         public string Facing { get; set; }
         public Nullable<int> Sft { get; set; }
         public Nullable<int> ClubHouseCharges { get; set; }
