+++ ModelLayer/Flats.cs
@@ -11,7 +11,7 @@ namespace ModelLayer
         public Nullable<int> ProjectID { get; set; }
         public Nullable<int> TowerID { get; set; }
         public Nullable<int> Floor { get; set; }
-        public Nullable<int> Bhk { get; set; }
+        public Nullable<double> Bhk { get; set; }
         public Nullable<int> Sft { get; set; }
         public string Facing { get; set; }
         public string BookingStatus { get; set; }
